requirements.txt
pandas
sqlalchemy
psycopg2-binary
python-dotenv
gspread
google-auth